+++ crescendo/data/_common.py
